Refactor into a class
